+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # unreleased
 
+## v0.1.0
+
+* doc: add some linux spcefic install instructions
+* feat: add darwin support w/brew
+* chore: ignore vagrant stuff for testing
+* feat: add mac signing script
+* fix: remove unused token attribute
+* feat: exit after --register or --inventory actions
+* fix: update to new structure layout for agent_type endpoint
+* chore: ignore .envrc
+* feat: name collectors -> agents
+* feat: rename agent id to manager id
+* feat: rename agent to manager
+* chore: update to go1.20.6 (GO-2023-1878)
+* feat: rename api endpoints/json manager registration
+* feat: chown configs to original gid/uid
+* chore(deps): bump golang.org/x/sys from 0.10.0 to 0.11.0
+* chore(deps): bump github.com/aws/aws-sdk-go-v2/config from 1.18.31 to 1.18.32
+
 ## v0.0.4
 
 * feat: add default api url